requirements.txt
# Core Dependencies
pandas==2.1.4
numpy==1.26.2

# Database - PostgreSQL
psycopg2-binary==2.9.9

# Config
pyyaml==6.0.1

# Optional: Für Ayoub (Feature 2) oder Adam (Dashboard)
# streamlit==1.29.0
# plotly==5.18.0
# scikit-learn==1.3.2

# Testing (optional)
# pytest==7.4.3
